Refactor Layakine audio scheduling into engine module
